requirements.txt
fastapi
langchain_core
langgraph
langchain_openai
python-dotenv
mlflow
matplotlib